Simplify the customization of ID generator.
Now P4_value.gen takes 2 optional arguments: one is the body of the ID generator which will be
inlined in the function definition; the other is a tuple value*type of a seed used by the ID
generator body, which will be passed to each recursive calls to value_of_* (for all the foreign
types).

The byproduct is that now, value_of_* should be constant (ie. the generated ID should all start
from the same number), and thus we can use = instead of Value.equal which was doing an equality
check modulo alpha-conversion of the IDs.